+++ rust/src/api/initialization.rs
@@ -1,8 +1,9 @@
-use crate::hardware_control::live_view::gphoto2;
-use crate::hardware_control::live_view::gphoto2::GPHOTO2_HANDLES;
+use crate::hardware_control::live_view::gphoto2::{self, GPHOTO2_HANDLES};
 use crate::hardware_control::live_view::nokhwa::NOKHWA_HANDLES;
 use crate::models::version_info::VersionInfo;
+use std::ffi::CStr;
 use std::sync::atomic::{Ordering, AtomicBool};
+use gexiv2_sys::gexiv2_get_version;
 pub use ipp::model::PrinterState;
 pub use ipp::model::JobState;
 use crate::{frb_generated::StreamSink, helpers::{self, HardwareInitializationFinishedEvent, TOKIO_RUNTIME}};
@@ -28,7 +29,29 @@ pub fn get_version_info() -> VersionInfo {
         rust_version: version().to_string(),
         rust_target: RUST_COMPILE_TARGET.to_owned(),
         library_version: LIBRARY_VERSION.to_owned(),
+        libgphoto2_version: ::gphoto2::library_version().unwrap().to_owned(),
+        libgexiv2_version: get_gexiv2_version(),
+        libexiv2_version: get_exiv2_version(),
     }
+}
+
+fn get_gexiv2_version() -> String {
+    let raw_version = unsafe { gexiv2_get_version() };
+    println!("{:?}", raw_version);
+
+    // Extract the major, minor, and patch versions
+    let major = raw_version / 10000;
+    let minor = (raw_version / 100) % 100;
+    let patch = raw_version % 100;
+
+    format!("{}.{}.{}", major, minor, patch)
+}
+
+fn get_exiv2_version() -> String {
+    return unsafe {
+        let c_str = crate::Exiv2_version();
+        CStr::from_ptr(c_str).to_string_lossy().into_owned()
+    };
 }
 
 pub fn initialize_hardware(ready_sink: StreamSink<HardwareInitializationFinishedEvent>) {