+++ rust/src/models/version_info.rs
@@ -3,4 +3,7 @@ pub struct VersionInfo {
     pub rust_version: String,
     pub rust_target: String,
     pub library_version: String,
+    pub libgphoto2_version: String,
+    pub libgexiv2_version: String,
+    pub libexiv2_version: String,
 }